split transformations from sphinxquery.cpp
